token_level_coinference/requirements.txt
transformers>=4.43.0
torch>=2.1.0
datasets>=2.18.0
bitsandbytes>=0.43.0
tqdm>=4.66.0
pyyaml>=6.0.1
numpy>=1.26.0